Added cast to avoid warnings of contrary anonymous enums in assignment
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1976 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/RenderOther/OverheadMapRenderer.cpp
@@ -183,7 +183,7 @@ void OverheadMapClass::Render(overhead_map_data& Control)
 				{
 					if (clockwise_polygon->floor_height!=counterclockwise_polygon->floor_height)
 					{
-						line_color= LINE_IS_LANDSCAPED(line) ? NONE : _elevation_line_color;
+						line_color= LINE_IS_LANDSCAPED(line) ? NONE : static_cast<short>(_elevation_line_color);
 					}
 				}
 			}
